# Optimize build process