--- a/.editorconfig
+++ b/.editorconfig
@@ -2,6 +2,6 @@
 root = true
 
 # Unix-style newlines with a newline ending every file
-[*.js]
+[*.js,*.ts]
 indent_style = space
 indent_size = 4