--- a/package.json
+++ b/package.json
@@ -16,7 +16,6 @@
     "@iobroker/js-controller-cli": "file:packages/cli",
     "@iobroker/js-controller-common": "file:packages/common",
     "@iobroker/js-controller-common-db": "file:packages/common-db",
-    "@iobroker/types": "file:packages/types",
     "iobroker.js-controller": "file:packages/controller"
   },
   "devDependencies": {
@@ -25,6 +24,8 @@
     "@alcalzone/release-script-plugin-lerna": "~3.5.0",
     "@alcalzone/release-script-plugin-license": "~3.5.0",
     "@alcalzone/release-script-plugin-manual-review": "~3.5.8",
+    "@iobroker/types-dev": "file:packages/types-dev",
+    "@iobroker/types": "file:packages/types-public",
     "@microsoft/api-extractor": "^7.33.2",
     "@tsconfig/node12": "^1.0.11",
     "@types/chai": "^4.3.3",
@@ -67,7 +68,6 @@
     "sinon": "^15.0.0",
     "sinon-chai": "^3.7.0",
     "ts-node": "^10.9.1",
-    "tsd": "^0.24.1",
     "typedoc": "^0.23.11",
     "typedoc-plugin-markdown": "^3.13.5",
     "typedoc-plugin-missing-exports": "^1.0.0",
@@ -90,6 +90,6 @@
     "test-jsonl": "mocha -r ts-node/register packages/controller/test/jsonl/*.ts --exit",
     "test-redis-socket": "mocha -r ts-node/register packages/controller/test/redis-socket/*.ts --exit",
     "test-redis-sentinel": "mocha -r ts-node/register packages/controller/test/redis-sentinel/*.ts --exit",
-    "test-types": "tsd packages/types/"
+    "test-types": "npm run test --workspace=@iobroker/types"
   }
 }