--- a/packages/adapter/tsconfig.json
+++ b/packages/adapter/tsconfig.json
@@ -4,6 +4,32 @@
     "outDir": "build",
     "rootDir": "src"
   },
+  "references": [
+    {
+      "path": "../db-objects-file/tsconfig.json"
+    },
+    {
+      "path": "../db-objects-jsonl/tsconfig.json"
+    },
+    {
+      "path": "../db-objects-redis/tsconfig.json"
+    },
+    {
+      "path": "../db-states-file/tsconfig.json"
+    },
+    {
+      "path": "../db-states-jsonl/tsconfig.json"
+    },
+    {
+      "path": "../db-states-redis/tsconfig.json"
+    },
+    {
+      "path": "../common/tsconfig.json"
+    },
+    {
+      "path": "../common-db/tsconfig.json"
+    },
+  ],
   "include": [
     "src/**/*.ts"
   ],