--- a/packages/cli/tsconfig.json
+++ b/packages/cli/tsconfig.json
@@ -4,6 +4,14 @@
     "outDir": "build",
     "rootDir": "src"
   },
+  "references": [
+    {
+      "path": "../common/tsconfig.json"
+    },
+    {
+      "path": "../common-db/tsconfig.json"
+    },
+  ],
   "include": [
     "src/**/*.ts",
     "src/**/*.js"