--- a/packages/common-db/tsconfig.build.json
+++ b/packages/common-db/tsconfig.build.json
@@ -10,6 +10,8 @@
     "src/**/*.ts"
   ],
   "exclude": [
-    "src/**/*.test.ts"
+    "src/**/*.test.ts",
+    "build/**",
+    "node_modules/**"
   ]
 }