--- a/packages/common-db/tsconfig.json
+++ b/packages/common-db/tsconfig.json
@@ -4,6 +4,29 @@
     "outDir": "build",
     "rootDir": "src"
   },
+  "references": [
+    {
+      "path": "../db-objects-file/tsconfig.json"
+    },
+    {
+      "path": "../db-objects-jsonl/tsconfig.json"
+    },
+    {
+      "path": "../db-objects-redis/tsconfig.json"
+    },
+    {
+      "path": "../db-states-file/tsconfig.json"
+    },
+    {
+      "path": "../db-states-jsonl/tsconfig.json"
+    },
+    {
+      "path": "../db-states-redis/tsconfig.json"
+    },
+    {
+      "path": "../common/tsconfig.json"
+    },
+  ],
   "include": [
     "src/**/*.ts"
   ],