--- a/packages/common/tsconfig.build.json
+++ b/packages/common/tsconfig.build.json
@@ -10,6 +10,8 @@
     "src/**/*.ts"
   ],
   "exclude": [
-    "src/**/*.test.ts"
+    "src/**/*.test.ts",
+    "build/**",
+    "node_modules/**"
   ]
 }