--- a/packages/controller/tsconfig.json
+++ b/packages/controller/tsconfig.json
@@ -4,6 +4,38 @@
     "outDir": "build",
     "rootDir": "src"
   },
+  "references": [
+    {
+      "path": "../db-objects-file/tsconfig.json"
+    },
+    {
+      "path": "../db-objects-jsonl/tsconfig.json"
+    },
+    {
+      "path": "../db-objects-redis/tsconfig.json"
+    },
+    {
+      "path": "../db-states-file/tsconfig.json"
+    },
+    {
+      "path": "../db-states-jsonl/tsconfig.json"
+    },
+    {
+      "path": "../db-states-redis/tsconfig.json"
+    },
+    {
+      "path": "../adapter/tsconfig.json"
+    },
+    {
+      "path": "../cli/tsconfig.json"
+    },
+    {
+      "path": "../common/tsconfig.json"
+    },
+    {
+      "path": "../common-db/tsconfig.json"
+    },
+  ],
   "include": [
     "src/**/*.ts",
     "src/**/*.js"