--- a/packages/db-base/tsconfig.build.json
+++ b/packages/db-base/tsconfig.build.json
@@ -6,6 +6,11 @@
     "checkJs": false,
     "noEmit": false
   },
+  "references": [
+    {
+      "path": "../common/tsconfig.build.json"
+    },
+  ],
   "include": [
     "src/**/*.ts",
     "src/**/*.js" // currently, we have js too