--- a/packages/db-objects-file/tsconfig.json
+++ b/packages/db-objects-file/tsconfig.json
@@ -4,6 +4,14 @@
     "outDir": "build",
     "rootDir": "src"
   },
+  "references": [
+    {
+      "path": "../db-base/tsconfig.json"
+    },
+    {
+      "path": "../db-objects-redis/tsconfig.json"
+    },
+  ],
   "include": [
     "src/**/*.ts"
   ],