--- a/packages/db-objects-redis/src/index.ts
+++ b/packages/db-objects-redis/src/index.ts
@@ -2,5 +2,5 @@ export function getDefaultPort(host: string): number {
     return host.includes(',') ? 26379 : 6379;
 }
 export const Server = null;
-export { ObjectsInRedisClient as Client, ChangeFileFunction } from './lib/objects/objectsInRedisClient';
+export { ObjectsInRedisClient as Client } from './lib/objects/objectsInRedisClient';
 export * as objectsUtils from './lib/objects/objectsUtils';