--- a/packages/db-objects-redis/tsconfig.json
+++ b/packages/db-objects-redis/tsconfig.json
@@ -4,6 +4,11 @@
     "outDir": "build",
     "rootDir": "src"
   },
+  "references": [
+    {
+      "path": "../db-base/tsconfig.json"
+    },
+  ],
   "include": [
     "src/**/*.ts",
     "src/**/*.js" // currently, we have js too