--- a/packages/db-states-file/tsconfig.build.json
+++ b/packages/db-states-file/tsconfig.build.json
@@ -11,6 +11,8 @@
     "src/**/*.js" // currently, we have js too
   ],
   "exclude": [
-    "src/**/*.test.ts"
+    "src/**/*.test.ts",
+    "build/**",
+    "node_modules/**"
   ]
 }