--- a/packages/db-states-file/tsconfig.build.json
+++ b/packages/db-states-file/tsconfig.build.json
@@ -6,6 +6,14 @@
     "checkJs": false,
     "noEmit": false
   },
+  "references": [
+    {
+      "path": "../db-base/tsconfig.build.json"
+    },
+    {
+      "path": "../db-states-redis/tsconfig.build.json"
+    },
+  ],
   "include": [
     "src/**/*.ts",
     "src/**/*.js" // currently, we have js too