--- a/packages/db-states-file/tsconfig.json
+++ b/packages/db-states-file/tsconfig.json
@@ -4,6 +4,14 @@
     "outDir": "build",
     "rootDir": "src"
   },
+  "references": [
+    {
+      "path": "../db-base/tsconfig.json"
+    },
+    {
+      "path": "../db-states-redis/tsconfig.json"
+    },
+  ],
   "include": [
     "src/**/*.ts"
   ],