--- a/packages/db-states-jsonl/tsconfig.build.json
+++ b/packages/db-states-jsonl/tsconfig.build.json
@@ -6,6 +6,17 @@
     "checkJs": false,
     "noEmit": false
   },
+  "references": [
+    {
+      "path": "../db-base/tsconfig.build.json"
+    },
+    {
+      "path": "../db-states-file/tsconfig.build.json"
+    },
+    {
+      "path": "../db-states-redis/tsconfig.build.json"
+    },
+  ],
   "include": [
     "src/**/*.ts",
     "src/**/*.js" // currently, we have js too