--- a/packages/db-states-redis/tsconfig.json
+++ b/packages/db-states-redis/tsconfig.json
@@ -4,6 +4,11 @@
     "outDir": "build",
     "rootDir": "src"
   },
+  "references": [
+    {
+      "path": "../db-base/tsconfig.json"
+    },
+  ],
   "include": [
     "src/**/*.ts"
   ],