--- a/packages/types-dev/package.json
+++ b/packages/types-dev/package.json
@@ -1,0 +1,30 @@
+{
+    "name": "@iobroker/types-dev",
+    "private": true,
+    "version": "5.0.0-alpha.0-20230217-a0c54039",
+    "engines": {
+        "node": ">=12.0.0"
+    },
+    "keywords": [
+        "ioBroker"
+    ],
+    "author": "foxriver76 <moritz.heusinger@gmail.com>",
+    "contributors": [
+        "foxriver76 <moritz.heusinger@gmail.com>"
+    ],
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/ioBroker/ioBroker.js-controller/packages/types-shared"
+    },
+    "main": "",
+    "types": "index.d.ts",
+    "license": "MIT",
+    "publishConfig": {
+        "access": "public"
+    },
+    "files": [
+        "index.d.ts",
+        "objects.d.ts",
+        "LICENSE"
+    ]
+}