--- a/packages/types-public/index.d.ts
+++ b/packages/types-public/index.d.ts
@@ -1,0 +1,2 @@
+import './build/shared';
+import './public';