--- a/packages/types-public/public.d.ts
+++ b/packages/types-public/public.d.ts
@@ -1,0 +1,11 @@
+// Types which are only safe to share publicly
+
+import { AdapterClass, AdapterOptions as _AdapterOptions } from './build/types';
+
+declare global {
+    namespace ioBroker {
+        type Adapter = AdapterClass;
+
+        type AdapterOptions = _AdapterOptions;
+    }
+}