--- a/tsconfig.json
+++ b/tsconfig.json
@@ -25,7 +25,7 @@
       "packages"
     ],
     "types": [
-      "@iobroker/types",
+      "@iobroker/types-dev",
       "node",
       "mocha"
     ],