… packages/types-dev/LICENSE
… packages/types-dev/objects.d.ts
… packages/types-public/api-extractor.json
… packages/types-public/index.test-d.ts
… packages/types/build.ts