The first set of changes brings the build time for JS-Controller down to 12s for a rebuild without changes, compared to ~1m10s previously, simply by disabling the prebuild script. Because we may need to force a clean build in some cases, I propose keeping the prebuild script around as `npm run clean`.

I'll investigate further with TS Project References.

**Edit:** While setting them up is tedious (effectively mirroring the dependencies into both tsconfig files per folder), they are **FAST**.
Down to 19s on a clean build (from 1m10s). Down to 1s on a rebuild (from 12s).