@AlCalzone 2023-02-19
@foxriver76 we have a circular dependency through packages, causing output to be treated as input files. All paths are relative to `packages/common`:
```
build/lib/common/tools.d.ts
  Imported via '@iobroker/js-controller-common/tools' from file '../db-objects-redis/build/lib/objects/objectsInRedisClient.d.ts' with packageId '@iobroker/js-controller-common/build/lib/common/tools.d.ts@5.0.0-alpha.0-20230217-a0c54039'
  Imported via '@iobroker/js-controller-common/tools' from file '../db-base/build/lib/inMemFileDB.d.ts' with packageId '@iobroker/js-controller-common/build/lib/common/tools.d.ts@5.0.0-alpha.0-20230217-a0c54039'
../db-base/build/lib/inMemFileDB.d.ts
  Imported via '@iobroker/db-base/inMemFileDB' from file '../db-objects-redis/build/lib/objects/objectsInRedisClient.d.ts' with packageId '@iobroker/db-base/build/lib/inMemFileDB.d.ts@5.0.0-alpha.0-20230217-a0c54039'
../db-objects-redis/build/lib/objects/objectsInRedisClient.d.ts
  Imported via './lib/objects/objectsInRedisClient' from file '../db-objects-redis/build/index.d.ts'
../db-objects-redis/build/index.d.ts
  Imported via '@iobroker/db-objects-redis' from file '../types/build/types.d.ts' with packageId '@iobroker/db-objects-redis/build/index.d.ts@5.0.0-alpha.0-20230217-a0c54039'
../types/build/types.d.ts
  Imported via './build/types' from file '../types/index.d.ts'
../types/index.d.ts
  Entry point of type library '@iobroker/types' specified in compilerOptions with packageId '@iobroker/types/index.d.ts@5.0.0-alpha.0-20230217-a0c54039'
```