@AlCalzone 2023-02-20
Solved it. @foxriver76 you may want to look at what I did in the last commit and keep it in mind for shared types you add in the future